==> documents/genomics/symcapture/22-GrowthAnimal.Rmd <==
```{r setup_growthanimal, include=FALSE}
rm(list = ls()) ; invisible(gc()) ; set.seed(42)
library(knitr)
library(kableExtra)
if(knitr:::is_html_output()) options(knitr.table.format = "html") 
if(knitr:::is_latex_output()) options(knitr.table.format = "latex") 
library(tidyverse)
library(raster)
library(bayesplot)
library(kinship2)
library(rstan)
theme_set(bayesplot::theme_default())
opts_chunk$set(
  echo = F, message = F, warning = F, fig.height = 6, fig.width = 8,
  cache = F, cache.lazy = F)
options(mc.cores = parallel::detectCores())
rstan_options(auto_write = T)
load("./symcapture_save/Env2.Rdata")
trees <- trees %>% 
  na.omit(pop) %>% # remove admixed
  dplyr::select(Ind, IID, pop, Plot, SubPlot, TreeFieldNum, Y0, DBH0, DBHtoday, N)
pairs_stan <- function(chain, stan_model, pars) {
  energy <- as.matrix(sapply(get_sampler_params(stan_model, inc_warmup = F), 
                             function(x) x[,"energy__"]))
  pars <- rstan::extract(stan_model, pars = pars, permuted = F)
  df <- data.frame(energy[,chain], pars[,chain,])
  names(df)[1] <- "energy"
  GGally::ggpairs(df, title = paste0("Chain", chain), 
                  lower = list(continuous = GGally::wrap("points", alpha = 0.2)))                    
} 
pars <- data.frame(sim = 1:25,
                   sigmaG = rep(seq(0.1, 1, length.out = 5), each = 5),
                   sgmaR = rep(seq(0.1, 1, length.out = 5), 5))
```

# Growth & Animal models

The aim of this document is to explore animal and growth model with generated data to validate their behavior and use it on *Symphonia* and *Eschweilera* real data. Let's consider a set of $P=3$ populations including each $Fam=3$ families composed of $I = 14$ individuals with arbitrar relationships (it's only 126 individuals to do quick tests).

```{r simkinship, fig.cap="Kinship matrix"}
simK <- function(
  P = 3, 
  Fam = 3
){
  I <- 14 # Individuals
  ped.df <- data.frame(pop = rep(1:P, each = Fam*I),
                       fam = rep(1:Fam, each = I),
                       ind = rep(1:I, Fam*P),
                       father = rep(c(NA, NA, 1, 1, 1, 1, NA, 
                                      6, 6, 6, NA, 11, 11, 11), Fam*P),
                       mother = rep(c(NA, NA, 2, 2, 2, 2, NA, 
                                      7, 7, 7, NA, 3, 3, 3), Fam*P), 
                       sex = rep(c(1, 2, 2, 3, 3, 1, 2, 3, 3, 
                                   3, 1, 3, 3, 3), Fam*P)) %>% 
    mutate_at(c("father", "mother"), funs(ifelse(!is.na(.), 
                                                 paste0(pop,fam,.), NA))) %>%
    mutate(ind = paste0(pop,fam,ind))
  ped.ped <- kinship2::pedigree(id = ped.df$ind, dadid = ped.df$father, 
                                momid = ped.df$mother, sex = ped.df$sex, 
                                famid = paste0(ped.df$pop, ped.df$fam))
  K <- as.matrix(kinship2::kinship(ped.ped))
  K <- 2*K
  return(list(df = ped.df, K = K))
}
heatmap(simK()$K)
```

## Animal

We used the following animal model with a lognormal distribution to estimate population and genotypic variance:

\begin{equation} 
  y_{p,i} \sim \mathcal{logN}(log(\mu_p.a_{i}),\sigma_1) \\
  a_{p,i} \sim \mathcal{MVlogN_N}(log(1),\sigma_2.K)
  (\#eq:animallogth)
\end{equation} 

```{r simanimal}
simAnimal <- function(
  P = 3, 
  Fam = 3,
  sigmaP = 0.5,
  sigmaG = 0.3,
  sigmaR = 0.2
){
  K <- simK(P, Fam)
  N <- nrow(K$K)
  epsilon <- rnorm(N)
  mu <- rlnorm(P, sd = sigmaP)
  y <- rlnorm(N, 
              log(mu[K$df$pop]) + sigmaG*as.vector(t(chol(K$K)) %*% epsilon), 
              sigmaR)
  mdata <- list(N = N, P = P, K = K$K, population = K$df$pop, y = y)
  return(list(mu = mu,
              Vp = var(log(mu[K$df$pop])), 
              Vg = var(sigmaG*as.vector(t(chol(K$K)) %*% epsilon)), 
              Vr = sigmaR^2,
              mdata = mdata))
}
```

We fitted the equivalent model with following priors:

\begin{equation} 
  y_{p,i} \sim \mathcal{logN}(log(\mu_p.\hat{a_{i}}), \sigma_1) \\ 
  \hat{a_{i}} = e^{\sqrt{V_G}.A.\epsilon_i} \\
  \epsilon_i \sim \mathcal{N}(0,1) \\ 
  ~ \\
  \mu_p \sim \mathcal{logN}(log(1),1) \\  
  \sigma_1 \sim \mathcal N_T(0,1) \\ 
  ~ \\
  V_Y = Var(log(y)) \\
  V_P = Var(log(\mu_p)) \\
  V_R=\sigma_1^2 \\
  V_G = V_Y - V_P - V_R \\
  (\#eq:animallogthstan)
\end{equation} 

```{r animalTable}
# mdata <- simAnimal(P = 3, Fam = 3)
# animal <- stan_model("symcapture_models/AnimalLog.stan")
# fitAnimal <- sampling(animal, chains = 2, save_warmup = F, data = mdata$mdata,
#                       init = "0",
#                       control = list(adapt_delta = 0.99, max_treedepth = 12))
# save(mdata, fitAnimal, file = file.path("symcapture_save", "animal.Rdata"))
load(file.path("symcapture_save", "animal.Rdata"))
broom::tidyMCMC(fitAnimal, pars = c("mu", "Vp", "Vg", "Vr", "lp__"), 
                droppars = NULL, rhat = T) %>% 
  mutate(expected = with(mdata, c(mu, Vp, Vg, Vr, NA))) %>% 
  dplyr::select(term, estimate, expected, std.error, rhat) %>% 
  kable(caption = "Animal model fitted versus expected values.",
        col.names = c("Parameter", "Estimate", "Expected",
                      "Standard error", "$\\hat R$"))
```

```{r animalTrace, fig.cap="Parameters for Animal model traceplot and expected value in red."}
mcmc_trace(as.array(fitAnimal, pars = c("mu", "Vp", "Vg", "Vr", "lp__")),
           np = nuts_params(fitAnimal)) +
  geom_hline(aes(yintercept = expected), col = "red", 
             data = data.frame(parameter = c(paste0("mu[", 1:length(mdata$mu), "]"), 
                                             "Vp", "Vg", "Vr"), 
                               expected = with(mdata, c(mu, Vp, Vg, Vr)))) 
```

## Gmax

We used the following growth model with a lognormal distribution to estimate individual growth potential:

\begin{equation} 
  DBH_{y=today,p,i}  - DBH_{y=y0,p,i} \sim \\
  \mathcal{logN} (log(\sum _{y=y_0} ^{y=today} \theta_{1,p,i}.exp(-\frac12.[\frac{log(\frac{DBH_{y,p,i}}{100.\theta_{2,p}})}{\theta_{3,p}}]^2)), \sigma_1) \\ 
  \theta_{1,p,i} \sim \mathcal {logN}(log(\theta_{1,p}), \sigma_2) \\ 
  \theta_{2,p} \sim \mathcal {logN}(log(\theta_2),\sigma_3) \\ 
  \theta_{3,p} \sim \mathcal {logN}(log(\theta_3),\sigma_4)
  (\#eq:gmaxth)
\end{equation} 


```{r simgmax}
simGmax <- function(
  P = 3, 
  I = 3*3*14,
  Y = 33, # Nb years
  Gmax = c(0.53, 0.54, 0.36),
  Dopt = 0.25,
  Ks = 0.7,
  sigma = c(0.2, 0.67, 0.2, 0.2)
){
  Gmax <- Gmax[1:P]
  years <- 1984:(1984+Y-1)
  pop <- sample(1:P, I, replace = T)
  Init <- sample_n(trees, I)
  DBH0 <- Init$DBH0
  Y0 <- Init$Y0
  Gmaxi <- rlnorm(I, meanlog = log(Gmax[pop]), sdlog = sigma[2])
  Doptp <- rlnorm(P, meanlog = log(Dopt), sdlog = sigma[3])
  Ksp <- rlnorm(P, meanlog = log(Ks), sdlog = sigma[4])
  DBH <- rep(NA, I)
  for (t in 1:(Y-1)) {
    for(i in 1:I){
      if(years[t] == Y0[i]) 
        DBH[i] <- DBH0[i] 
    }
    DBH <- DBH + Gmaxi * exp(-0.5*(log(DBH / (100*Doptp[pop])) / Ksp[pop])^2)
  }
  DBH <- DBH - DBH0
  DBHtoday <- DBH0 + rlnorm(I, meanlog = log(DBH), sdlog = sigma[1])
  return(list(Gmax = Gmax, Dopt = Dopt, Ks = Ks, 
              sigmaR = sigma[2], sigma = sigma[1],
              Vr = sigma[2]^2, Vp = var(log(Gmax[pop])),
              mdata = list(I = I, Y = Y, P = P, years = years, DBH0 = DBH0,
                           Y0 = Y0, DBHtoday = DBHtoday, pop = pop)))
}
```

We fitted the equivalent model with following priors:

\begin{equation} 
  DBH_{y=today,p,i}  - DBH_{y=y0,p,i} \sim \\
  \mathcal{logN} (log(\sum _{y=y_0} ^{y=today} \hat{\theta_{1,p,i}}.exp(-\frac12.[\frac{log(\frac{DBH_{y,p,i}}{100.\hat{\theta_{2,p}}})}{\hat{\theta_{3,p}}}]^2)), \sigma_1) \\ 
  \hat{\theta_{1,p,i}} = e^{log(\theta_{1,p}) + \sigma_2.\epsilon_{1,i}} \\ 
  \hat{\theta_{2,p}}  = e^{log(\theta_2) + \sigma_3.\epsilon_{2,p}} \\ 
  \hat{\theta_{3,p}}  = e^{log(\theta_3) + \sigma_4.\epsilon_{3,p}} \\  
  \epsilon_{1,i} \sim \mathcal{N}(0,1) \\
  \epsilon_{2,p} \sim \mathcal{N}(0,1) \\
  \epsilon_{3,p} \sim \mathcal{N}(0,1) \\
  ~ \\
  (\theta_{1,p}, \theta_2, \theta_3) \sim \mathcal{logN}^3(log(1),1) \\ 
  (\sigma_1, \sigma_2, \sigma_3, \sigma_4) \sim \mathcal{N}^4_T(0,1) \\ 
    ~ \\
  V_P = Var(log(\mu_p)) \\ 
  V_R=\sigma_2^2
  (\#eq:gmaxthstan)
\end{equation} 

```{r gmaxTable}
# mdata <- simGmax()
# gmax <- stan_model("symcapture_models/Gmax.stan")
# fitGmax <- sampling(gmax, chains = 2, save_warmup = F, data = mdata$mdata,
#                     control = list(adapt_delta = 0.99, max_treedepth = 12))
# save(mdata, fitGmax, file = file.path("symcapture_save", "gmax.Rdata"))
load(file.path("symcapture_save", "gmax.Rdata"))
broom::tidyMCMC(fitGmax, pars = c("thetap1", "theta2", "theta3",
                                  "Vp", "Vr", "lp__"), 
                droppars = NULL, rhat = T) %>% 
  mutate(expected = with(mdata, c(Gmax, Dopt, Ks,  
                                  Vp, Vr, NA))) %>% 
  dplyr::select(term, estimate, std.error, expected, rhat) %>% 
  kable(caption = "Animal model fitted versus expected values.",
        col.names = c("Parameter", "Estimate", "Standard error", 
                      "Expected", "$\\hat R$"))
```

```{r gmaxTrace, fig.cap="Parameters for Growth model traceplot and expected value in red."}
mcmc_trace(as.array(fitGmax, pars = c("thetap1", "theta2", "theta3",
                                      "Vp", "Vr", "lp__")),
           np = nuts_params(fitGmax)) +
  geom_hline(aes(yintercept = expected), col = "red", 
             data = data.frame(parameter = c(paste0("thetap1[", 
                                                    1:length(mdata$Gmax), "]"),
                                             "theta2", "theta3", "Vp", "Vr"), 
                               expected = with(mdata, c(Gmax, Dopt, Ks,  Vp, Vr))))
```

## Gmax & Animal

```{r simgmaxgeno}
simGmaxGeno <- function(
  P = 3, 
  Fam = 3,
  Y = 33, # Nb years
  Gmax = c(0.53, 0.54, 0.36),
  Dopt = 0.25,
  Ks = 0.7,
  sigma = c(0.20, 0.67, 0.2, 0.2, 0.4)
){
  K <- simK(P, Fam)
  I <- nrow(K$K)
  Gmax <- Gmax[1:P]
  years <- 1984:(1984+Y-1)
  pop <- K$df$pop
  Init <- sample_n(trees, I)
  DBH0 <- Init$DBH0
  Y0 <- Init$Y0
  epsilonG <- rnorm(I)
  a <- exp(log(Gmax[pop]) + sigma[5]*as.vector(t(chol(K$K)) %*% epsilonG))
  Gmaxi <- rlnorm(I, meanlog = log(a), sdlog = sigma[2])
  Doptp <- rlnorm(P, meanlog = log(Dopt), sdlog = sigma[3])
  Ksp <- rlnorm(P, meanlog = log(Ks), sdlog = sigma[4])
  DBH <- rep(NA, I)
  for (t in 1:(Y-1)) {
    for(i in 1:I){
      if(years[t] == Y0[i]) 
        DBH[i] <- DBH0[i] 
    }
    DBH <- DBH + Gmaxi * exp(-0.5*(log(DBH / (100*Doptp[pop])) / Ksp[pop])^2)
  }
  DBH <- DBH - DBH0
  DBHtoday <- DBH0 + rlnorm(I, meanlog = log(DBH), sdlog = sigma[1])
  return(list(Gmax = Gmax, Dopt = Dopt, Ks = Ks, 
              sigmaR = sigma[2], sigmaG = sigma[5], sigma = sigma[1],
              Vp = var(log(Gmax[pop])), 
              Vg = var(sigma[5]*as.vector(t(chol(K$K)) %*% epsilonG)), 
              Vr = sigma[2]^2, 
              mdata = list(I = I, Y = Y, P = P, years = years, DBH0 = DBH0,
                           Y0 = Y0, DBHtoday = DBHtoday, pop = pop, K = K$K)))
}
```

We used the following growth model with a lognormal distribution to estimate individual growth potential and associated genotypic variation:

\begin{equation} 
  DBH_{y=today,p,i}  - DBH_{y=y0,p,i} \sim \\
  \mathcal{logN} (log(\sum _{y=y_0} ^{y=today} \theta_{1,p,i}.exp(-\frac12.[\frac{log(\frac{DBH_{y,p,i}}{100.\theta_{2,p}})}{\theta_{3,p}}]^2)), \sigma_1) \\ 
  \theta_{1,p,i} \sim \mathcal {logN}(log(\theta_{1,p}.a_{1,i}), \sigma_2) \\ 
  \theta_{2,p} \sim \mathcal {logN}(log(\theta_2),\sigma_3) \\ 
  \theta_{3,p} \sim \mathcal {logN}(log(\theta_3),\sigma_4) \\
  a_{1,i} \sim \mathcal{MVlogN}(log(1), \sigma_5.K)
  (\#eq:gmaxgenoth)
\end{equation} 

We fitted the equivalent model with following priors:

\begin{equation} 
  DBH_{y=today,p,i}  - DBH_{y=y0,p,i} \sim \\
  \mathcal{logN} (log(\sum _{y=y_0} ^{y=today} \hat{\theta_{1,p,i}}.exp(-\frac12.[\frac{log(\frac{DBH_{y,p,i}}{100.\hat{\theta_{2,p}}})}{\hat{\theta_{3,p}}}]^2)), \sigma_1) \\ 
  \hat{\theta_{1,p,i}} = e^{log(\theta_{1,p}.\hat{a_{1,i}}) + \sigma_2.\epsilon_{1,i}} \\ 
  \hat{\theta_{2,p}}  = e^{log(\theta_2) + \sigma_3.\epsilon_{2,p}} \\ 
  \hat{\theta_{3,p}}  = e^{log(\theta_3) + \sigma_4.\epsilon_{3,p}} \\  
  \hat{a_{1,i}} = e^{\sigma_5.A.\epsilon_{4,i}} \\
  \epsilon_{1,i} \sim \mathcal{N}(0,1) \\
  \epsilon_{2,p} \sim \mathcal{N}(0,1) \\
  \epsilon_{3,p} \sim \mathcal{N}(0,1) \\
  \epsilon_{4,i} \sim \mathcal{N}(0,1) \\
  ~ \\
  (\theta_{1,p}, \theta_2, \theta_3) \sim \mathcal{logN}^3(log(1),1) \\ 
  (\sigma_1, \sigma_2, \sigma_3, \sigma_4, \sigma_5) \sim \mathcal{N}^5_T(0,1) \\ 
    ~ \\ 
    V_P = Var(log(\mu_p)) \\
    V_G=\sigma_5^2\\
    V_R=\sigma_2^2
  (\#eq:gmaxgenothstan)
\end{equation} 

```{r gmaxgenoTable}
# mdata <- simGmaxGeno()
# gmaxgeno <- stan_model("symcapture_models/GmaxGeno.stan")
# fitGmaxGeno <- sampling(gmaxgeno, chains = 2, save_warmup = F, data = mdata$mdata,
#                         control = list(adapt_delta = 0.99, max_treedepth = 12))
# save(mdata, fitGmaxGeno, file = file.path("symcapture_save", "gmaxgeno.Rdata"))
load(file.path("symcapture_save", "gmaxgeno.Rdata"))
broom::tidyMCMC(fitGmaxGeno, pars = c("thetap1", "theta2", "theta3",
                                      "Vp", "Vg", "Vr", "lp__"), 
                droppars = NULL, rhat = T) %>% 
  mutate(expected = with(mdata, c(Gmax, Dopt, Ks,  
                                  Vp, Vg, Vr, NA))) %>% 
  dplyr::select(term, estimate, std.error, expected, rhat) %>% 
  kable(caption = "Animal model fitted versus expected values.",
        col.names = c("Parameter", "Estimate", "Standard error", 
                      "Expected", "$\\hat R$"))
```

```{r gmaxgenoTrace, fig.cap="Parameters for Growth & Animal model traceplot and expected value in red."}
mcmc_trace(as.array(fitGmaxGeno, pars = c("thetap1", "theta2", "theta3",
                                          "Vp", "Vg", "Vr", "lp__")), 
           np = nuts_params(fitGmaxGeno)) +
  geom_hline(aes(yintercept = expected), col = "red", 
                          data = data.frame(parameter = c(paste0("thetap1[", 
                                                    1:length(mdata$Gmax), "]"),
                                             "theta2", "theta3", "Vp", "Vg", "Vr"), 
                               expected = with(mdata, c(Gmax, Dopt, Ks, Vp, Vg, Vr))))
```
